blog/api/permissions.py
from rest_framework import permissions
from blog.models import Article

class IsOwnerOrReadonly(permissions.BasePermission):
    def has_object_permission(self, request, view, obj):
        if request.method in permissions.SAFE_METHODS:
            return True
    
        return obj.author==request.user